Fixing Receiver issue
Previously, when the receiver got a duplicate start packet, it created a new
connection object, and since we keep track of connections by the IP address and
port tuple, it mistakenly erased the state it had for a connection, causing
incorrect acks to be sent. This fix causes the sender to never create a new
connection objects for an (IP,port) for which we have an active connection.