Add test framework
Feedback from last year said that we didn't provide enough guidance on
testing.  Also, the autograder code was a hot mess. As a result, the
evaluation process was much less transparent than we'd like it to be.

To rectify this, this year we're adding a basic test harness with the
project. The idea is to use the exact same test harness for both
grading and for students' own testing. We provide a test harness that
acts as a packet forwarder between a sender and receiver
implementation. Students (or GSIs...) can write tests that plug into
this forwarder, allowing them to perform arbitrary operations on the
packet stream. We provide students with a basic "test" that just
passes all packets through the forwarder, and a second test that
randomly drops packets. Hopefully this will allow us to write cleaner
tests, and let students run some basic tests on their own code before
they submit it.