Fix handling of checksums in update_packet and _send
Because we were re-calculating the checksum every time we called _send in the
forwarder, it wasn't possible to corrupt a packet. We now do not update the
checksum in send.

In addition, we now calculate the checksum in update_packet over the
non-zero-indexed sequence number, which is necessary due to the above change.
Also fix a bug that was causing packets to be incorrectly re-written when their
data section was empty (this issue was being masked by the checksum
recalculation in _send).